Update README.md
When implementing this in my own projects I had problems with the import line but worked when using const instead. 

import { ColorPaletteControl } = gumponents.components;
const { PostRelationshipControl } = gumponents.components;

So when implementing these example lines of code in a .js and .jsx file I git a vscode error stating that it expects 'from' instead of '=' on the import line. To be extensive it expects a string literal in a .ts file as well.

So in the README.md file I have changed import to const which works in my code.